refactor: Remove Room persistence from Pitch
Initially, my intention was to persist a table of Pitch
entities, which would contain the pitch information for all
chromatic pitches given in the string array resource in
strings.xml.

After further consideration, Pitch.java has been redesigned
to instantiate this String list of pitch names and cache
them in the pitchNames arraylist.  Functions were previously
added to support generating a frequency value from a string
pitch name.  With this in place, there is no need to have the
pitch persistence, and all references to this have been removed.