test: Add additional tests for AndroneDao
insertAndroneReplacesOnConflict() - ensure inserted Androne
sharing an ID with an existing Androne replaces the existing
Androne.

insertAndroneAndGetList() - insert a new Androne and
ensure call for List<Androne> returns a list containing
that Androne.

updateAndroneAndGetById() - insert androne, update the
existing object, update the dao, get the androne back
from the Dao and ensure it contains the updated values.

All tests pass to completion as of the time of this commit