refactor: Remove pitchListIndex member variable
Having the pitchListIndex member variable increasing the coupling
of the Androne object to the Pitch object, since the Androne was
handling logic of whether or not pitch/frequency will be in between
min/max ranges allowed.  Replaced this with calls to new Pitch
increment/decrement functions, which handle all of the logic
within the Pitch class.  From the pitch class, throw
IllegalArgumentExceptions when the increment/decrement call is
not possible.