refactor: Redesign Androne's setFrequency and setPitchName methods
Originally, these two methods would create a new pitch object with
the argument passed into the methods.  Restructure it so that
it instead calls pitch's setFrequency and setPitchName methods,
freeing the need to create a new pitch object every time the
user wants to change the frequency or pitch name.

Add error handling in cases where pitch or frequency set is out of
the allowable range