refactor: Combine stopAndrone and playAndrone to togglePlay
Having a stopAndrone and playAndrone method forces the view to
consider the state of the Androne, whether or not it is playing.
By combining this into a single method, togglePlay(), this
allows the presenter to handle the logic of whether or not to
call the Androne object's stop / play methods.